unify sparse vector fixtures
