+++ lib/sparse/Cargo.toml
@@ -18,6 +18,4 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tempfile = "3.8.0"
 ordered-float = "4.1"
-
-[dev-dependencies]
 rand = "0.8.5"
