+++ lib/sparse/src/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod sparse_vector;
+pub mod sparse_vector_fixture;
 pub mod types;
